requirements.txt
chess-graph==1.7.5
pgnparser==1.0
plotly==4.7.1
python-chess==0.31.1
python-lichess==0.9
requests==2.23.0
retrying==1.3.3
urllib3==1.25.9
chardet==3.0.4
six==1.15.0